moves the open in discover equation table fields logic to a util function called getEquationFields so it can be reused. This will be useful for the widget viewer
