+++ static/app/utils/discover/fields.tsx
@@ -1,5 +1,6 @@
 import isEqual from 'lodash/isEqual';
 
+import {parseArithmetic} from 'sentry/components/arithmeticInput/parser';
 import {RELEASE_ADOPTION_STAGES} from 'sentry/constants';
 import {Organization, SelectValue} from 'sentry/types';
 import {assert} from 'sentry/types/utils';
@@ -1196,4 +1197,25 @@ export function hasDuplicate(columnList: Column[], column: Column): boolean {
     return false;
   }
   return columnList.filter(newColumn => isEqual(newColumn, column)).length > 1;
+}
+
+export function getEquationFields(fields: string[]): string[] {
+  // Gather all fields and functions used in equations and prepend them to the provided fields
+  const fieldsClone = [...fields];
+  const termsSet: Set<string> = new Set();
+  // Gets equation terms
+  fieldsClone.forEach(field => {
+    if (isEquation(field)) {
+      const parsed = parseArithmetic(stripEquationPrefix(field)).tc;
+      parsed.fields.forEach(({term}) => termsSet.add(term as string));
+      parsed.functions.forEach(({term}) => termsSet.add(term as string));
+    }
+  });
+  // Pushes terms to returned fields
+  termsSet.forEach(term => {
+    if (Array.isArray(fieldsClone) && !fieldsClone.includes(term)) {
+      fieldsClone.unshift(term);
+    }
+  });
+  return fieldsClone;
 }
